backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic>=1.18.0
asyncpg==0.29.0
pgvector==0.2.4

# LLM & Embeddings
dashscope==1.17.0
openai>=1.0.0

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
pdfplumber==0.10.3
pymupdf>=1.23.0
pillow>=10.0.0

# Text Processing
jieba==0.42.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
loguru==0.7.2

# Authentication
bcrypt==4.1.2
PyJWT==2.8.0
cryptography>=42.0.0
resend>=2.0.0

# Validation & Serialization
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1